requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
authlib
tenacity
aiolimiter
openai
